Add Wordfence WAF template and installation task
The WAF file is used by Wordfence Extended Protection.
https://www.wordfence.com/help/firewall/optimizing-the-firewall/

Normally, the WAF file is generated by Wordfence on the dashboard but it
requires user interaction. This template is used to generate an
equivalent file which is automatically installed to the given path using
the task in this commit.

The variable 'wordfence_waf_path' does not have a default and has to be
explicitly defined.

Note: The template assumes that Wordfence is installed as a must-use
plugin (in 'mu-plugins' directory).